+++ src/main.cpp
@@ -26,7 +26,7 @@
 #include <sentry.h>
 #endif
 
-#define LSP_VERSION "1.46.0"
+#define LSP_VERSION "1.47.0"
 
 static void displayFlags()
 {