+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [1.47.0] - 2025-05-17
+
 ### Added
 
 - Implemented opt-in crash reporting for the language server. This is enabled via `luau-lsp.server.crashReporting.enabled`. Note: crash reporting